new: extend new plugin system
